+++ storedimo/database/migrations/2025_06_11_145041_add_db_host_to_empresas.php
@@ -13,9 +13,15 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('empresas', function (Blueprint $table) {
-            $table->text('db_host')->nullable()->after('id_tipo_bd'); // db host
-        });
+        // Schema::table('empresas', function (Blueprint $table) {
+        //     $table->text('db_host')->nullable()->after('id_tipo_bd'); // db host
+        // });
+
+        if (Schema::hasTable('empresas') && !Schema::hasColumn('empresas', 'db_host')) {
+            Schema::table('empresas', function (Blueprint $table) {
+                $table->text('db_host')->nullable()->after('id_tipo_bd'); // db host
+            });
+        }
     }
 
     /**
@@ -25,8 +31,14 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('empresas', function (Blueprint $table) {
-            $table->dropColumn('db_host'); // db host
-        });
+        // Schema::table('empresas', function (Blueprint $table) {
+        //     $table->dropColumn('db_host'); // db host
+        // });
+
+        if (Schema::hasTable('empresas') && Schema::hasColumn('empresas', 'db_host')) {
+            Schema::table('empresas', function (Blueprint $table) {
+                $table->dropColumn('db_host'); // db host
+            });
+        }
     }
 };
